Fixed problems
