+++ app/src/main/java/com/example/nicocalculator/Utils/ConstsUtils.kt
@@ -1,0 +1,16 @@
+package com.example.nicocalculator.Utils
+
+const val DECIMAL_SEPARATOR = '.'
+const val VALUE_MIN_CAR = 48
+const val VALUE_MAX_CHAR = 57
+const val VALUE_SIZE_OPERATION = 3
+const val VALUE_EMPTY_TEXT = ""
+const val VALUE_ZERO = 0
+const val ADD_OPERATOR = "+"
+const val SUBTRACT_OPERATOR = "-"
+const val MULTIPLY_OPERATOR = "*"
+const val DIVIDE_OPERATOR = "/"
+const val OPEN_PAR = "("
+const val CLOSE_PAR = ")"
+const val OPERATIONS_DEFAULT_TEXT = "Operations"
+const val RESULTS_DEFAULT_TEXT = "Results"
